Alternative solution for problem 2 using a Map
Point lookup is done with an extension to the mathematics used for
problem 1 rather than generating a point stream and using that for
lookup.

The rest of the solution to the problem is remarkably simple when done
like this. I really did overengineer my first solution in the goal of
handling arbitrarily sized input.